requirements.txt
streamlit
langchain
langchain-groq
chromadb
pandas
openpyxl
python-dotenv
tiktoken
sentence_transformers
protobuf<5
pysqlite3-binary